Move config api content
